* tests/tcltest.test: Added 'exit' calls to the scripts that the test creates and added tests for -tmpdir and -args.
* tests/env.test:
* tests/exec.test: Added 'exit' calls to scripts that the tests
themselves write.

* tests/README: Cleaned up documentation in README so that it
refers to variables and documentation that actually exist.  Also
added missing documentation for flag usage.

* library/tcltest1.0/tcltest.tcl: Added a -args flag that sets a
variable named ::tcltest::parameters based on whatever's being
sent in as the argument to the -args flag.  Modified tcltest so
that tcltest namespace variables are only initialized to their
default values if they did not previously exist.  Modified the
::tcltest::testConstraints variable so that it isn't unset every
time ::tcltest::initConstraints is called.  Modified command line
processing so that they are only processed if ::tcltest doesn't
have a child namespace (if there is a child namespace, then the
command line arguments *must* be processed in that namespace).